Changed pwm variable to a double so small changes can be kept

+++ micro_ros_raspberrypi_pico_sdk/hardware/Motor.cpp
@@ -43,7 +43,7 @@ int Motor::getCurrentPwm(){
     return currentPwm;
 }
 
-void Motor::setSpeed(int speedPWM) {
+void Motor::setSpeed(double speedPWM) {
     // Ensure PWM is within the valid range
     if (currentPwm == speedPWM) {
         return;
